Termiando el diseño de la vista recuperar contraseña

+++ src/front/js/store/flux.js
@@ -72,7 +72,7 @@ const getState = ({
                     })
                     .catch((err) => console.log(err));
             },
-            
+
             login: (userEmail, userPassword) => {
                 fetch(
                         "https://3001-agusalvs-salvandopatita-aarew83zv9d.ws-us85.gitpod.io/api/autenticacion", {
@@ -125,6 +125,46 @@ const getState = ({
                                     popup: "animate__animated animate__fadeOutUp",
                                 },
                             });
+                        }
+                        localStorage.setItem("token", data.access_token);
+                    })
+                    .catch((err) => console.log(err));
+            },
+
+            recuperar: (userNuevacontraseña, userRepetircontraseña) => {
+                fetch(
+                        "https://3001-agusalvs-salvandopatita-yp2yoipd64w.ws-us85.gitpod.io/api/recuperar", {
+                            method: "POST",
+                            headers: {
+                                "Content-Type": "application/json",
+                                // 'Content-Type': 'application/x-www-form-urlencoded',
+                            },
+                            body: JSON.stringify({
+                                nuevacontraseña: userNuevacontraseña,
+                                repetircontraseña: userRepetircontraseña,
+                            }), // body data type must match "Content-Type" header
+                        }
+                    )
+                    .then((response) => {
+                        console.log(response.status);
+                        if (response.status === 201) {
+                            Swal.fire({
+                                position: "middle",
+                                icon: "success",
+                                title: "Contraseña cambiada correctamente",
+                                showConfirmButton: false,
+                                timer: 1500,
+                            });
+                            setStore({
+                                auth: true,
+                            });
+                        }
+                        return response.json();
+                    })
+                    .then((data) => {
+                        console.log(data);
+                        if (data.msg === "Bad email or password") {
+                            alert(data.msg);
                         }
                         localStorage.setItem("token", data.access_token);
                     })
